distinguish compiler and linker flags
The '-mthumb-interwork' and '-march=' flags
are related to the code-generation process
but not to the link process.

In general case this options are platform-related,
but ML-supported cameras are not very different,
so we can set this options in one place.

Anyway, there is no need to write '-march=armv5te'
in multiple places.

--HG--
branch : unified